requirements.txt
pydexcom==0.2.3
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.7.2
numpy==1.24.3
scipy==1.11.1
pytest==7.4.0
pytest-mock==3.11.1